requested changes
- merge methods to param arrays
- move StructureJigsawPoolInitEvent  to separate class
- remove constructor with ResourceLocation in SingleJigsawPiece
